+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project are documented in this file.
 ## Unreleased
 
  - Skia renderer: Fixed the `source` property of `Image` elements sometimes not changing when setting dynamically loaded images. (#3510)
+ - Fixed native style not clipping correctly GroupBox (#3541)
 ## [1.2.1] - 2023-09-19
 
 ### General